requirements.txt
arff==0.9
cycler==0.11.0
et-xmlfile==1.1.0
gitdb==4.0.9
gitpython==3.1.18
joblib==1.1.0
kiwisolver==1.3.2
lizard==1.17.9
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
pillow==9.0.0
pydriller==2.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
scikit-fuzzy==0.4.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
six==1.16.0
smmap==5.0.0
threadpoolctl==3.0.0
requests==2.27.1
bs4
pyyaml